misc/xwalk-system-tests/tests/navigatorlanguage/res/w3c/language.html

<!--
postMessage(navigator.language);
/*
-->
<!doctype html>
<title>navigator.language</title>
<script src="../../../../js/w3c/testharness.js"></script>
<script src="../../../../js/w3c/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var worker = new Worker('#');
  worker.onmessage = this.step_func(function(e) {
    assert_equals(e.data, navigator.language);
    this.done();
  });
});
</script>
<!--
*/
//-->
